requirements.txt
requests
Pillow
spacy
scikit-learn
numpy
